requirements.txt
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
idna==2.8
isort==4.3.17
lazy-object-proxy==1.3.1
mccabe==0.6.1
more-itertools==7.0.0
pkg-resources==0.0.0
pluggy==0.9.0
py==1.8.0
pylint==2.3.1
pytest==4.4.1
PyYAML==5.1
requests==2.21.0
six==1.12.0
slackclient==1.3.1
typed-ast==1.3.4
urllib3==1.24.2
websocket-client==0.54.0
wrapt==1.11.1
